move query invitation info to admin api for RLS
