+++ lib/supabase/apis/invitation.ts
@@ -13,22 +13,6 @@ export const createInvitation = async (creator_id: number, group_id: number) => 
   return invitation as InvitationInfo;
 };
 
-export const getInvitationInfo = async (code: string) => {
-  if (code.length !== INVITATION_CODE_LENGTH) throw 'invalid code';
-
-  const minCreatedTime = addDays(new Date(), -1);
-  const { data, error } = await supabase
-    .from('invitations')
-    .select(`*, groups ( name ), users ( nickname )`)
-    .eq('code', code)
-    // @note: 생성시점이 min(= 현재 - 24시간) 보다 커야 유효한 시점
-    .gt('created_time', minCreatedTime.toISOString());
-  const invitationInfo = data?.[0];
-
-  if (error) throw error;
-  return invitationInfo as InvitationInfo | undefined;
-};
-
 export const getInvitationsByUserId = async (creator_id: number, group_id: number) => {
   const minCreatedTime = addDays(new Date(), -1);
   const { data, error } = await supabase
